Feat: reduce unsafe code in verifier

+++ foaks/prime_field/src/lib.rs
@@ -10,12 +10,12 @@ use std::{
 
 use self::error::{PrimeFieldError, RootOfUnityError};
 
-pub const MOD: u64 = 2305843009213693951;
+pub const MOD: usize = 2305843009213693951;
 
 pub static mut INITIALIZED: AtomicBool = AtomicBool::new(false);
 
 pub const MASK: u32 = 4294967295; // 2^32 - 1
-pub const PRIME: u64 = 2305843009213693951; // 2^61 - 1
+pub const PRIME: usize = 2305843009213693951; // 2^61 - 1
 
 pub const MAX_ORDER: usize = 62;
 
@@ -40,11 +40,11 @@ impl FieldElementContext {
     }
 }
 
-pub fn my_mod(x: u64) -> u64 {
+pub fn my_mod(x: usize) -> usize {
     (x >> 61) + (x & MOD)
 }
 
-pub fn my_mult(x: u64, y: u64) -> u64 {
+pub fn my_mult(x: usize, y: usize) -> usize {
     // return a value between [0, 2PRIME) = x * y mod PRIME
     // return ((hi << 3) | (lo >> 61)) + (lo & PRIME)
     let (lo, hi) = x.widening_mul(y);
@@ -92,8 +92,8 @@ impl VecFieldElement {
 
 #[derive(Serialize, Default, Debug, PartialEq, Eq, Copy, Clone)]
 pub struct FieldElement {
-    pub real: u64,
-    pub img: u64,
+    pub real: usize,
+    pub img: usize,
 }
 
 impl FieldElement {
@@ -110,29 +110,29 @@ impl FieldElement {
     }
 
     pub fn new_random() -> Self {
-        let real = rand::random::<u64>() % MOD;
-        let img = rand::random::<u64>() % MOD;
+        let real = rand::random::<usize>() % MOD;
+        let img = rand::random::<usize>() % MOD;
 
         Self::new(real, img)
     }
 
     pub fn new_random_real_only() -> Self {
-        let real = rand::random::<u64>() % MOD;
+        let real = rand::random::<usize>() % MOD;
 
         Self::new(real, 0)
     }
 
-    pub const fn from_real(real: u64) -> Self {
+    pub const fn from_real(real: usize) -> Self {
         let real = real % MOD;
         Self { img: 0, real }
     }
 
-    pub const fn from_img(img: u64) -> Self {
+    pub const fn from_img(img: usize) -> Self {
         let img = img % MOD;
         Self { img, real: 0 }
     }
 
-    pub const fn new(real: u64, img: u64) -> Self {
+    pub const fn new(real: usize, img: usize) -> Self {
         Self { img, real }
     }
 
@@ -144,7 +144,7 @@ impl FieldElement {
         Self::new(1, 0)
     }
 
-    pub fn sum_parts(&self) -> u64 {
+    pub fn sum_parts(&self) -> usize {
         self.real + self.img
     }
 
@@ -187,14 +187,14 @@ impl FieldElement {
     //}
 }
 
-fn verify_lt_mod_once(mut a: u64) -> u64 {
+fn verify_lt_mod_once(mut a: usize) -> usize {
     if a >= MOD {
         a -= MOD;
     }
     a
 }
 
-fn verify_lt_mod_many(mut a: u64) -> u64 {
+fn verify_lt_mod_many(mut a: usize) -> usize {
     while a >= MOD {
         a -= MOD;
     }
